requirements.txt
alembic==1.8.1
anyio==3.6.1
asttokens==2.0.8
autopep8==1.7.0
backcall==0.2.0
bcrypt==4.0.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
debugpy==1.6.3
decorator==5.1.1
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
entrypoints==0.4
executing==1.1.0
fastapi==0.85.0
greenlet==1.1.3
h11==0.14.0
httptools==0.5.0
idna==3.4
ipykernel==6.16.0
ipython==8.5.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter-core==4.11.1
jupyter_client==7.3.5
Mako==1.2.3
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.5
orjson==3.8.0
packaging==21.3
parso==0.8.3
passlib==1.7.4
pickleshare==0.7.5
prompt-toolkit==3.0.31
psutil==5.9.2
psycopg2==2.9.3
pure-eval==0.2.2
pyasn1==0.4.8
pycodestyle==2.9.1
pycparser==2.21
pydantic==1.10.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
ujson==5.5.0
urllib3==1.26.12
uvicorn==0.18.3
watchfiles==0.17.0
wcwidth==0.2.5
websockets==10.3
windows-curses==2.3.0